[icon-aes:master] Squashed commit of [icon-aes:new-thermo] (B. Stevens), MR(icon/icon-aes!56):
Corrections to the moist thermodynamics were introduced through

- changes to the saturation adjustment scheme to conserve internal energy and delete unnecessary code
- small modifications to the graupel scheme to correct the temperature tendencies
- modification to the calling sequence of physical routines within mo_aes_phys_main.f90 so microphysics is called first

These necessitated a rewrite of the old mo_satad.f90 scheme, which resulted in considerable simplification,
and the introduction of a local aes version of graupel to avoid premature negotiation about the structure
of this scheme with the rest of the ICON community.

The new thermodynamics change the results (hopefully for the better as they address energy leaks in the old
formulation). They have been tested with the bubble case, both by visually inspecting the results and
evaluating the change in the internal energy -- the latter being a more critical test. Changes were not
tested for cases with ice processes substantially active (the bubble produces trace amounts of ice in its
present configuration) and should be tested using a full atmosphere configuration before being accepted.
Further test cases need to be clarified to facilitate more continuous integration and development.