[cimd] [icon-cimd] fix to get correct time written into mf restart patches
https://gitlab.dkrz.de/icon/icon-cimd/-/merge_requests/89

The time variable as it is written to the multifile restart patch files
with format "day as %Y%m%d.%f" requires double precision upon writing
and during the calculation. Otherwise, truncation leads to wrong time
values.

Merged-by: Rene Redler <rene.redler@mpimet.mpg.de>

Squashed commit of the following:

* [icon-cimd] fix to get correct time written into mf restart patches

The time variable as it is written to the multifile restart patch files
with format "day as %Y%m%d.%f" requires double precision upon writing
and during the calculation. Otherwise, truncation leads to wrong time
values.