[cscs] Suggestion: Acc beautifier CI
https://gitlab.dkrz.de/icon/icon-cscs/-/merge_requests/54

This would add a CI Pipline that checks the style of OpenACC Code in ICON.


You might have noticed that we - the GPU community - unified and cleaned up the OpenACC code in ICON two weeks ago (24c5f97e0c326726456539be3a719487f960453f). This included a unification of indentation, upper/lower case, spaces, commas, unification of the lacc-switch, removal of obsolete ifdefs and more.
We oriented our work along the ICON OpenACC style and implementation guide and amended the guide according to our work.
Part of the code changes was done manually and part of them was done automatically using an ACC [beautifier script](https://gitlab.dkrz.de/dwd-sw/icon-openacc-beautifier) that I wrote.
The beautifier can check and correct the ACC style within ACC directives in typical ICON code.
To keep the uniformity that we achieved during the clean-up effort, we would like to introduce a continuous integration (CI) pipeline based on the beautifier.
With Jfe's help, I was able to implement such pipeline in a Gitlab branch.
The pipeline runs on every commit to a merge request and gives a green/red indication whether the code complies with the ACC style guide or not.
In addition, the pipeline provides a patch file, which can be applied directly with git, when the code does not follow the style guide.
This should help the programmer to adjust her code easily and has the great potential to skip all discussions on ACC code style in the review.

Please checkout my example MR which includes some intentional violations of the ACC style guide:
https://gitlab.dkrz.de/icon/icon-cscs/-/merge_requests/55
You should note the red X from the latest pipeline, the corresponding error message and attached artifacts (which will be deleted automatically by next Tuesday).
The pipeline only takes a few seconds, i.e. it is much lighter than buildbot.
I configured it, such that it runs automatically on each commit to a MR (if you push to a branch that is not yet part of a MR, nothing will happen, but you should be able to trigger the pipeline manually in this case.)

Approved-by: William Sawyer <wsawyer@cscs.ch>
Merged-by: William Sawyer <wsawyer@cscs.ch>

Squashed commit of the following:

* fix one more path

* fix path

* Use fixed and public revision.

* Upload patch as artifact on failure

* try acc_style.patch

* no body

* CI_MERGE_REQUEST_PROJECT_ID

* try rest api

* try rules

* rename job

* remove -e "\n"

* fix a typo

* improve user friendliness

* cd ..

* try python3

* use CI-token

* try url directly

* change url to https

* Try acc beautifier as CI