[icon-mpim:master] Use HD mask from JSBACH
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/19

Changes provided by Rene Redler.

For the aes physics with JSBACH we now use the mask array for the discharge as provided 
by the hd_para file and thus mark only those cells as discharge cells which carry values. 
Previously, all ocean cells were marked as discharge cells with the majority carrying zeros. 
While the old implementation was not wrong and delivered valid results it does not work 
anymore with very high resolutions of 2.5 or 1.2 km as the amount of MPI messages YAC would 
generate during the initial search for the source-to-target-mapping would explode.

Merged-by: Reiner Schnur <reiner.schnur@mpimet.mpg.de>

Squashed commit of the following:

* [feature-land-hd-mask] Some additional cleanup
  * destruct_atmo_coupling not needed anymore
  * jsb_fdef_hd_fields not used in mo_aes_coupling, deleted

* [feature-land-hd-mask] Minor fix

* [feature-land-hd-mask] Use HD mask for yac from JSBACH (part 2)

  This commit provides the necessary changes in JSBACH ...
  jsbach.git:feature/yac-hd-mask@4a4d2f29

* [feature-land-hd-mask] Use HD mask for yac from JSBACH

  This commit needs a corresponding change in JSBACH!