Sppt port
See merge request icon/icon-nwp!764

Gpu port of the recently added sppt functionality.

New references are needed for the sppt experiment due to a small namelist change in radiation (ecrad_isolver=2 so the radiation runs on GPU).

This MR also includes some QOL changes that are not directly needed for sppt and are open for discussion:
1. removal of some obsolete code considering a negative nproma in icon-dev.checksuite
2. the addition of 5 development serialization points
3. Renaming of the sppt_namelist from `sppt_for_domain_` to `sppt_of_domain_` to be inline with the other namelist names

Approved-by: Xavier Lapillonne <xavier.lapillonne@meteoswiss.ch>
Approved-by: Marek Jacob <marek.jacob@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>