GPU Port of subroutines relevant for the instantaneous SST-ICE mode
See merge request icon/icon-nwp!760

The subroutine `process_sst_and_ice` was ported to OpenACC. However, only the branch for `lseaice=.FALSE.` (i.e. no sea ice) was ported, since the branch for sea ice could not be tested with the experiment created in MR !759. This experiment is over the Gulf of Genoa (the choice of domain was limited to the Alps domain, due to the lack of lateral BC elsewhere) where there is no ice.

In more details, the following was done:
- Ported the initialization of `t_seasfc` and `fr_ice` before the call to `process_sst_and_ice`.
- Ported the main loop for the case where the sea ice is switched off (`lseaice=.FALSE.`).
- Ported `aggregate_tg_qvs`.
- Added `lacc` flag where needed to switch between CPU and GPU.
- Added stop for the `lseaice=.TRUE.` branch.

_Note: ~~this MR should be merged together with MR !759, which adds the setup for the new test case.~~ --> I merged MR !759 in this branch already._

Merged-by: Daniel Reinert <daniel.reinert@dwd.de>