[mpim] Move configuration of the bundled library to the build stage
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/31

This introduces a new configure option `--enable-delayed-config` (defaults to `yes`). When the option is disabled (i.e. `--disable-delayed-config`), everything should work as it does now (or very close to it, at least). Otherwise (i.e. `--enable-delayed-config`), the configuration of the bundled libraries takes place at the build time (i.e. when running `make`). The advantage of that is that the libraries can be configured in parallel when running `make` with more than one job (e.g. `make -j8`). This makes the configuration stage about 5-7 times faster while the build stage takes 5-10% more time. The total configuration plus build time on Levante drops from ~17min to ~15min.

Merged-by: Sergey Kosukhin <sergey.kosukhin@mpimet.mpg.de>