name_list_output: fix non-advancing timestamps when appending to an output file across restart
See merge request icon/icon-nwp!787

When appending to an output file after restart, the appended time steps all have the same time stamp. This happened because of a recent fix for a double-destroy error, where the CDI time axis object was duplicated from the CDI stream opened for appending. The new time stamps are written to the duplicate and read by CDI from the original, which never gets updated.

This changeset removes some of the special-case code for appending files. It removes the caching of the IDs associated with the CDI stream object, and copying IDs back and forth between two sets of variables. When the time axis or vlist ID of the current file is needed it is retrieved from the CDI stream.

Approved-by: Florian Prill <florian.prill@dwd.de>
Merged-by: Florian Prill <florian.prill@dwd.de>