iterative IAU by looping over perform_nh_stepping
See merge request icon/icon-nwp!769

In an attempt to reduce the complexity of `mo_nh_stepping` and to improve readability, 
large parts of the code for controlling the iterative IAU has been removed from `mo_nh_stepping`.

The iterative IAU is now implemented at a higher level by calling `perform_nh_stepping` twice with different model stop dates `time_config%tc_stopdate`. 
For the first iteration, the stop date is set to `time_config%tc_exp_startdate`, such that the model integration stops after `dt_shift` (i.e. the time by which the actual start date is shifted ahead of the nominal date). For the second iteration, the stop date is set to its original user-defined value.
The save/restore routines `save_initial_state`, `restore_initial_state` are called directly before the first and second call of `perform_nh_stepping`, respectively.

As a result, the manual reset of dates and control switches (like `mtime_current`, `linit_dyn`, ...) as well as the wrapper routine `reset_to_initial_state` could be removed from the code in large parts.

What is left, unfortunately, is the necessity to reset all non-constant code elements that are created or initialized prior to the call of `perform_nh_stepping`, but which are modified thereafter. In particular, this applies to

* time level arrays `nnow`, `nnew`, `nnow_rcf`, `nnew_rcf` 
* mtime events (for NWP physics, DACE, EMVORADO, LHN)
* prognostic and diagnostic model fields (handled by the save/restore mechanism)

#### Differences between the first and second iteration

The following points are treated differently during the first and second IAU iteration. Note that this has not changed in comparison to the original implementation. It is documented here for completeness:

* the first iteration `iau_iter=1` uses an asymmetric IAU window, such that the analysis increments are completely added until the nominal start date `time_config%tc_exp_startdate`. The second iteration `iau_iter=2` uses the standard symmetric IAU window.
* the call of `update_nwp_phy_bcs` is skipped for `iau_iter=1`. This routine updates several slowly varying surface variables at midnight, such as albedo, ndviratio, etc. By not calling `update_nwp_phy_bcs` we avoid the necessity to reset those fields.
* **Output**
  * output at the nominal start date `time_config%tc_exp_startdate` is only written at the end of the first cycle (`iau_iter=1`) and skipped for the second cycle (`iau_iter=2`). By this mechanism, we provide an initialized analysis to which the analysis increments have been completely added.
  * Meteograms are handled in an analogous manner: Meteogram sampling at the nominal start date is only performed during the first cycle, and skipped for the second cycle (see call of `meteogram_sample_vars`). However, I am not 100% sure if this makes sense for meteograms.
* The initialization of several physics packaged is skipped for `iau_iter=2` (see `lreset=.FALSE.` in `init_nwp_phy`). In particular this is the case for
  * upper-atmospher physics
  * vege_clim
  * ecRad
   
  This complication might be avoided by calling the corresponding destruct or cleanup routines at the end of `perform_nh_stepping` (problem with ecRad: no destruct routine available for `ecrad_conf`)
* The radar forward operator EMVORADO is skipped for `iau_iter=1`, with the exception of the initial time step.
* The DACE observation operator is called during "negative simulation times" for the first cycle (`iau_iter=1`). For the second cycle, however, it starts at the nominal start date (i.e. for `sim_time>0`).
* prefetching of boundary data is skipped at the nominal start date for the first cycle (`iau_iter=1`).

Merged-by: Daniel Reinert <daniel.reinert@dwd.de>