rework nwp ocean interface to be usable with TERRA/turbtran and JSBACH/VDIFF
See merge request icon/icon-nwp!567

The interface relies on tile-based fields for surface variables and fluxes that are filled by TERRA/turbtran but unavailable in JSBACH/VDIFF. To build a unified interface to the coupler, this changeset introduces two structures of pointers `t_nwp_ocean_fields_tx` and `t_nwp_ocean_fields_rx` for the fields that are sent to and received from the coupler by the atmosphere.

Depending on context, the pointers can be set to the tile-based fields or to fields present in VDIFF's data structures. A wrapper function with the old name `nwp_couple_ocean` is provided, doing exactly that. Use of the wrapper produces bit-identical results compared to the previous implementation. However, the wrapper slightly changes the semantics of the input flux averaging. Instead of using `frac_t(isub_water)` and `frac_t(isub_seaice)` as weights directly, it calculates a sea-ice fraction `frac_si = frac_t(isub_seaice)/(frac_t(isub_water) + frac_t(isub_seaice))`, where the denominator is the ocean fraction of the grid box, and uses `frac_si` and `(1 - frac_si)`. This only has an effect on grid points with fractional ocean cover, but should give correct results there while the previous formula underestimates the fluxes in that case.

There are some open questions:

* Instead of `frac_si`, one could use the sea-ice fraction `diag_lnd%fr_seaice` directly, which should give the area of ice per area of ocean. Why wasn't this done?
* The previous implementation had provisions for adding hail, graupel, and ice to the snow flux, but those were disabled. The changeset contains a few lines where these fields are added opportunistically when available (i.e., when the field pointers in `prm_diag` are associated). Because they break bit-identity, they are commented out. Should they be included or not? **Including them is fine** (Martin, 2023-04-04)
* Are the boundary exchanges through `sync_patch_array` really necessary? None of the fields are used in a context where halo cells are needed, as far as I can see.

@m211032 @b380439

Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>