[mpim] Revision of the interface structure for the AES physics
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/8

The goal is to simplify the control structure and the code of the interfaces used to couple the dynamics and the parameterizations.

Some control switches are removed (lparamcpl, ldrymoist) and their default implementation is kept.
Removal of several unused/unnecessary variables, rewrite comments, change the order of computations.
Shift call of "two" cloud microphysics, cloud droplet number and WMO tropopause diagnostics.
Clean up of code. 

Author: Marco Giorgetta <marco.giorgetta@mpimet.mpg.de>
Co-authored-by: Monika Esch <monika.esch@mpimet.mpg.de>
Co-authored-by: Sergey Kosukhin <sergey.kosukhin@mpimet.mpg.de>
Co-authored-by: Daniel Reinert <daniel.reinert@dwd.de>
Co-authored-by: Daniel Hupp <daniel.hupp@meteoswiss.ch>
Merged-by: Monika Esch <monika.esch@mpimet.mpg.de>