Fix 700 error in advection
See merge request icon/icon-nwp!843

There is an atomic update (increment by 1) in a kernel (VECTOR_LENGTH=128) in `mo_advection_traj.f90` on a gang private variable. This means this variable should only be at most 128. Since the update is in an `if` statement it could also be smaller. However, I observed in some cases the variable to be 129 or even 130. Doing an atomic increment without a condition in that kernel adds up to 256, even though `NVCOMPILER_ACC_NOTIFY=1` confirms the VECTOR_LENGTH is 128.

Update: This happens because the maximum number of gangs allowed in the compiler is 65536, so in this code the assumption that each vector does only one loop iteration is wrong. Because we try to do more than 65536 gangs, the compiler only creates 65536 gangs and has to let some vectors run multiple loop iterations. WIth this fix we make sure that `gang_eidx` and `gang_elev` are more than big enough.

Merged-by: Daniel Reinert <daniel.reinert@dwd.de>