[mpim] Replaced CLAW directives and preprocessing with pure OpenACC
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/92

Replaced CLAW directives and preprocessing with pure OpenACC directives in ICON-Land/JSBACH
and remove CLAW from the ICON build system. (There are also three files in ICON containing
CLAW directives)

Also, the external `externals/omni-xmod-pool` has been removed.

With this change, it is best to configure and compile in a clean directory!

Approved-by: Sergey Kosukhin <sergey.kosukhin@mpimet.mpg.de>
Merged-by: Reiner Schnur <reiner.schnur@mpimet.mpg.de>