[mpim] Bugfix for carbon conservation test
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/116

Bugfix for carbon conservation test

See MR jsbach/jsbach!89

Parameter options%nc was not set in calculate_current_c_ta_state_sum, instead initial
value -1 was used. In simulations with carbon, this led to a crash with the nag compiler
and to wrong results of the carbon conservation test otherwise.

Approved-by: Reiner Schnur <reiner.schnur@mpimet.mpg.de>
Merged-by: Veronika Gayler <veronika.gayler@mpimet.mpg.de>