Option to modify diagnostic cloud scheme to enhance cloud cover in stratocumulus regions
See merge request icon/icon-nwp!881

This branch uses the Bechtold and Marquet stability criterion (EIS) together with a new inversion height diagnostic to identify regions containing stratocumulus. Cloud cover in these regions can be enhanced by modifying the diagnostic relationship between humidity and clc from quadratic to linear. The exponent of the relationship is scaled linearly between 2 (default) and 1 (stratocumulus) for temperatures dropping below -5C down to -15C, and also scales linearly with EIS at the edge of the identified region to allow a smoother transition.
This change leads to improvements in radiation RMSE vs. CERES, and T850 scores.
This branch is NOT bit identical with master, even if this enhancement option is disabled, because the exponent in the cloud cover relationship is now a float (vs. integer previously).

Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>