New namelist parameter for specifying the CFL-W threshold for adaptive time step reduction
See merge request icon/icon-nwp!890

A new namelist parameter named `vcfl_threshold` is introduced that allows specifying the CFL-W threshold at which the adaptive reduction of the dynamics time step (increase of `ndyn_substeps`) is triggered. Selecting a value less than the default of 1.05 is recommended for global resolutions between 5 and 10 km if occasional numerical instabilities occur otherwise; this is considered to be a more efficient alternative to a general reduction of the time step.

Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>