Horizontal mean for local grids (LAM)
See merge request icon/icon-nwp!788

The function Sum_2D_2Dweights_InRange (shared/mo_statistics.f90) uses the grid subset "owned" (all cells except halo cells) to calculate the horizontal mean. This function is (amongst others) used by SR levels_horizontal_mean called from within mo_nwp_diagnosis to calculate horizontal means ("global means", e.g. "tas_gmean"). For LAM (or for a subdomain when nesting is used), the lateral boundary interpolation zone has to be excluded from this calculation (for a detailed description of a treatment of the different boundary zones, see the icon-nwp tutorial https://doi.org/10.5676/dwd_pub/nwv/icon_tutorial2020), particularly Fig. 8.2.

Therefore, a new SUBROUTINE fill_subset_indices is called instead of the SR fill_subset to create the subset "owned_no_boundary", which is now used as an argument in the call to "levels_horizontal_mean" instead of the "standard" subset "owned".

Additionally, a test (run/checksuite.clm/exp.ICON_CLM_global_mean_no_boundary.run) is provided which shows that the boundary interpolation zone is excluded now.
Therefore, all cells of the temporary array used for the horizontal averaging are filled with a HUGE value in the boundary interpolation zone, so that a comparison of the mean value calculated by cdo fldmean in the post-processing with the *_geman value can easily detect differences. The test automatically detects if the result is incorrect. For a correct functioning of the test, lmask_boundary in io_nml has to be set to .true.

The modifications were not tested with nesting or with a global grid.

Approved-by: Daniel Rieger <daniel.rieger@dwd.de>
Approved-by: Ralf Mueller <ralf.mueller@dkrz.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>