[hotfix] fix reference data creation on levante
https://gitlab.dkrz.de/icon/icon/-/merge_requests/196

Reference data creation is currently done with the `update` test mode of the `icon-dev.checkuite` script. This leads to errors because this script is generating and modifying runscripts without taking into account the runtime options (provided by `--runflags`) in the experiment lists. Hence falling back to defaults, which might contradict to the initial test purpose.

This MR simplifies this workflow by splitting reference data creation and comparison (both is done with the `update` test mode and an appended copy job).

A new experiment list called `update-ref` is responsible for the data creation,only. It simply runs the plain tests and store the data in the new directory. A double-triggering the the `update` test list is no longer needed.

Merged-by: Ralf Mueller <ralf.mueller@dkrz.de>