Port Hailcast to GPU
See merge request icon/icon-nwp!897

**Integrate hailcast model from WRF in ICON**

Integrate a modified version, which includes GPU port and some adaptations and fixes from the WRF model.
This modified verison was originally used in the COSMO model. 

**Additional output variables**
- dhail_mx (maximum hailsize over all 5 sample hail stones,  reset intervall 1H)
- dhail_av (average hailsize over the 5 sample hail stones,  reset intervall 1H)
- dhail_sd (standard deviation of hail size over the 5 sample hail stones, reset intervall 1H)

Currently only dhail_mx has a correct valid grib2 coding and can be used in nwp output.

All the variables above are based on field _dhail_, which is computed in hailcast_driver and holds
the _MAX_ values of each species (or sample hail stones) since the last rest (to zero).

**The field _dhail_ has a reset intervall of 1H **

**Additonal namelist variables** __io_nml__
- dt_hailcast (calling frequency [seconds] of hailcast for dhail_* diagnostics, default 3min
- wdur_min_hailcast (minimal updraft persistence [seconds] for hailcast to be activated, 1H intervall) default 15 min

**Testcase**  
The additonal fields are added to _exp.mch_opr_r19b07_lpi_.  

wdur_min is supposed to be used for testing only, since the time-intervall of 2-5 mins of the testcases
is simply to short for a real case activation of hailcast.

**Timings on GPU**

Run on 3 nodes with 3 GPUs on Piz Daint (single processor run not possible because of data exchange in hailcast).

dt_hail: 20s  
wdur_min: 10s  
dt: 10s  
10 min total simulation  
output frequency: 60s  
**With hailcast (optional_diagnostics_atmoshphere): 0.00080s (0.009s)**  
**With hailcast (optional_diagnostics_atmoshphere): 0.00058s (0.006s)**

Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Florian Prill <florian.prill@dwd.de>