[seamless] fix if statement in lsm_ocean_atmo
See merge request icon/icon-nwp!958

Trang @b364148 found a mistake in the routine lsm_ocean_atmo where <1.0 was replaced with <10.0.  This cascade of IF ELSE statement was such that only values <=1.0 were left at this point.  We expect no significant difference from this fix. 
This is also interesting to Stephan @m211032.

Merged-by: Daniel Reinert <daniel.reinert@dwd.de>