[mpim] Fix GitLab CI scripts
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/172

This MR fixes false positive results of the checks introduced in
icon/icon-mpim!170. The logic of the scripts relied on their relative path with
respect to the root source directory of ICON, which has not been adjusted after
they were moved to a different directory. This MR fixes the bug in the scripts
and makes them more explicit and robust.

Merged-by: Sergey Kosukhin <sergey.kosukhin@mpimet.mpg.de>