real_timer: allow differing subtimers between PEs in aggregated report
See merge request icon/icon-nwp!973

This is required for timing inside conditionally called parts of ICON,
e.g. for terra when there are PEs without land points. The current
behavior excludes timers that are not activated by all PEs from the
report. The routine should also generate an error message, but that
does not happen on NEC because the condition for printing the message
is wrong (`my_process_is_stdio` is false for all VE PEs).

This changeset computes the union of subtimers over all participating
PEs, creating a common global timer hierarchy. Timers whose parents
differ across PEs are moved to the root, same as timers that are
activated locally from multiple contexts. Since the hierarchy is built
in `get_timer_reductions` and the consistency checks are gone, all
communication and a lot of complexity can be removed from the printing
routine `print_report_hierch_agg`.

To keep consistent counts and averages, the artificial minimum of one
timer activation per PE is removed (times are only ever divided by that
number if it is nonzero because the timer has to be activated at least
once before it is considered "used" and included in the report). We also
keep the number of PEs that activated a timer at least once in a new
field of `t_timer_reductions` for use in computing the average total
time. The total time is set to `HUGE(1._dp)` for the minimum reduction
on PEs that did not use a timer, so the value does not interfere with
timings from PEs that actually used it. For the other statistics it is
sufficient to keep the default value of zero.

Approved-by: Florian Prill <florian.prill@dwd.de>
Merged-by: Florian Prill <florian.prill@dwd.de>