[mpim] Add MPI-M GPU configure wrapper
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/179

This MR adds a configure wrapper for the GPU configuration on MPI-M machines.

Since there are no machines at MPI-M where we could run the GPU code, the
compiler flags set by the wrapper do not specify any particular compute
capability version as in the wrappers for real GPU systems. Additionally, the
compiler is requested to translate OpenACC directives not only to the GPU code
but also to the multicore CPU code (see comments in the wrapper on how to run
it). The wrapper is mainly meant for testing of the GPU-related build system
features.

This MR also updates other MPI-M wrappers in several ways:
- '-D__SWAPDIM' is removed for all CPU configurations;
- '-DOCE_SOLVE_OMP' is moved to 'ICON_OCEAN_FCFLAGS' wherever possible;
- compiler flags for the NVHPC CPU wrapper are copied from icon/icon-dkrz!64.

Merged-by: Sergey Kosukhin <sergey.kosukhin@mpimet.mpg.de>