[mpim] Refactor the NVIDIA/PGI function inlining
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/180

This MR refactors and applies minor fixes to the NVIDIA/PGI function inlining
implemented in the build system:

1. The list of source files that contain the inlined functions undergoes
additional filtering: files that are not needed for the current configuration
are not processed by the compiler during the inline library generation. For
example, RTE-RRTMGP-related files do not have to be masked anymore with
@RTE_RRTMGP_ENABLED@ in inlib.mk.in: they will be filtered out anyway when ICON
is configured with '--disable-rte-rrtmgp'.

2. The list of inlined functions is passed to the compiler not only when the
inline library is generated but also when it is used. First, this ensures that
functions that are not currently listed in the pgi_inlib_procedures variable but
(for whatever reason) appear in the inline library will not be inlined. Second,
this seems to save a number of redundant attempts to read the TOC file of the
inline library. The assumption is that the compiler does not try to read the
contents of the inline library if none of the functions in the current
translation unit is mentioned on the command line.

3. The GPU register limiting flag is now applied also when any of the extra
preprocessing steps is enabled (e.g. '--enable-serialization') and only if the
GPU support is enabled (i.e. '--enable-gpu'). The pattern- and target-specific
assignments of the compiler flags are also rearranged to make sure that the
flags do not propagate to the prerequisites (see comments in icon.mk.in).

Merged-by: Sergey Kosukhin <sergey.kosukhin@mpimet.mpg.de>