[mpim] Add CMake wrapper to support bundled libraries with complex dependencies
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/169

This MR adds a wrapper that can be used for the integration of CMake-based
bundled libraries. The wrapper is supposed to help in the following cases:

1. A bundled library is not under the control of the ICON community and,
therefore, we cannot ask the developers of the library to introduce
ICON-specific modifications into their CMake scripts.

2. A bundled library has a dependency that requires a complex set of linker
flags to be passed to the Fortran compiler when linking the ICON executable.

For more details, see the MR description.

Merged-by: Sergey Kosukhin <sergey.kosukhin@mpimet.mpg.de>