[nwp] Initial ART port and port of the pollen-related processes to GPU
See merge request icon/icon-nwp!756

This branch contains the initial changes to the ICON code that are necessary to run ICON on GPU with ART enabled, and to carry out the port of the pollen related processes that are contained in the ART submodule.

A wiki page exists for the planning of the ART port: https://gitlab.dkrz.de/icon/wiki/-/wikis/General-information-about-ART-on-GPU

Initially, this branch is tested using the MCH experiment: `mch_opr_r04b07_lhn_12`, the relevant ART input data is in:
`/scratch/e1000/meteoswiss/scratch/gvanpary/icon.pollen/gpu/ART_testcase` on Balfrin. Don't forget to add the `+art` in the spack command used to build ICON. In the runscript, the following namelists are added:
(make sure to change the path to redirect the input to your folder). 
```
! in run_nml
 lart                        =                      .TRUE.        ! ICON-ART main switch

...

&art_nml
 lart_aerosol         = .true.
 iart_volcano         = 0
 iart_seasalt         = 0
 iart_radioact        = 0   
 iart_dust            = 0
 iart_pollen          = 1
 lart_diag_out        = .true.
 iart_init_aero       = 0
 cart_input_folder    = '/scratch/e1000/meteoswiss/scratch/gvanpary/icon.pollen/gpu/ART_testcase'
 cart_modes_xml       = '/scratch/e1000/meteoswiss/scratch/gvanpary/icon.pollen/gpu/ART_testcase/input/xml_nl/modes.xml'
 cart_aerosol_xml     = '/scratch/e1000/meteoswiss/scratch/gvanpary/icon.pollen/gpu/ART_testcase/input/xml_nl/tracers.xml'
 cart_diagnostics_xml = '/scratch/e1000/meteoswiss/scratch/gvanpary/icon.pollen/gpu/ART_testcase/input/xml_nl/diagnostics.xml'
 lart_chem            = .false.
 lart_chemtracer      = .false.
 iart_aci_warm        = 0
 iart_aci_cold        = 0
 iart_ari             = 0          ! [0|1] for radiation feedback with ART aerosols (ASH or DUST)
 iart_aero_washout    = 1
 lart_pntSrc          = .false.
 cart_pntSrc_xml      ='/scratch/e1000/meteoswiss/scratch/gvanpary/icon.pollen/gpu/ART_testcase/input/data/pntSrc.xml'
/

&output_nml
 output_filename  = "ART"
 filename_format  = "<output_filename>_<datetime2>"
 filetype         = 5
 remap            = 0
 output_grid      = .TRUE.
 output_bounds    = ${output_bounds}        ! start, end, increment
 steps_per_file   = ${steps_per_file}
 include_last     = .FALSE.
 ml_varlist       =
  'group:ART_AEROSOL', 'group:ART_ROUTINE_DIAG'
/
```
Here is a list of changes that the first commit contains, which were made in order to have a running experiment on GPU with ART enabled:
- In the ART aerosol dynamics interface: the subroutine art_aerodyn_interface is manually set to CPU by putting the flag `i_am_accel_node` during the subroutine call.
- In the ART sedimentation interface: for now commented is the call to the GPU version of the subroutine `upwind_vflux_ppm` which is `upwind_vflux_ppm4gpu`, this needs to be adressed in further commits.
- In the nonhydrostatic timestepping: the ART subroutines are called with the `i_am_accel_node` flag set to `.FALSE.`.
- The subroutine `WMO_tropopause` is extended to be callable both on CPU and on GPU depending on the optional `lacc` flag.
- The various calls to `FINISH` when `lart = .TRUE.` for a GPU run are removed. 

Note that currently the output contains only 0s and in order to fix and continue the port, the ART code needs to be modified by somoene with Developer rights to the ART repository.

In the COSMO code, the equivalent port and openACC statements are in the files `organize_pollen.f90` and `pol_emissions.f90`.

Please ensure that the relevant people are kept up to date about the advances on this project:
- Sven Werchner (KIT, ART code maintainer, @b380678),
- Andreas Pauling (MCH, pollen processes, @b381983),
- Marek Jacob (DWD, OpenACC porting, @b380572).


Note that according to [issue 9474](https://gitlab.dkrz.de/icon/icon/-/issues/9474), MCH ports the pollen code on the branch art-openacc of this merge request. However, not the whole port is contained in icon-nwp/art-openacc. The remaining pollen port is done in the art repository (on the branch art/art-nwp). The corresponding merge request is [!54](https://gitlab.dkrz.de/art/art/-/merge_requests/54).

Approved-by: Daniel Hupp <daniel.hupp@meteoswiss.ch>
Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>