[nwp] mch opr tests on bb
See merge request icon/icon-nwp!972

This adds a new list to BB (create_list_mch-opr). In this list are tests of the operational setup at MeteoSwiss. Currently there are only three performance tests, but we might extend this list in the future.

These tests need more time than our usual CI/CD tests and so this list will not be triggered for each MR but will be set up to be run weekly once this MR is merged. Produced performance data will be stored permanently for possible later analysis.

For the time limit of the BB slurm jobs, I added some logic that, although not perfect, improves upon the previous version.

Approved-by: Daniel Hupp <daniel.hupp@meteoswiss.ch>
Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>