Revision of NWP buildbot tests
See merge request icon/icon-nwp!1000

The DWD/NEC buildbot tests are revised in order to reduce their consumption of compute resources, with the goal to include the whole set of tests in the standard icon-dev list. The previous extended test suite `list_dwd` is removed in turn.

Experiment 04, which used to test one-way nesting combined with starting and terminating a nest during runtime, is merged into Exp. 05

Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>