[nwp] Add missing ASYNC(1) clause to PARALLEL regions
See merge request icon/icon-nwp!954

Some parts of the source code still involve `ACC PARALLEL` regions without the `ASYNC(1)` clause.

This may lead to problems if the appropriate `ACC WAIT` statements are missing. For example, the CPU may send a kernel to the `ASYNC(1)` queue, then continue its computations, before sending a new (synchronous) kernel to the GPU. If the first kernel is still running, this may lead to errors.

We thus add the `ASYNC(1)` clause to every `ACC PARALLEL` region in the `src/` directory. The `ACC WAIT(1)` statements are added with safety in mind, according to the [style guide](https://gitlab.dkrz.de/icon/wiki/-/wikis/GPU-development/ICON-OpenAcc-style-and-implementation-guide#use-of-asynchronous-kernel-execution).

For the benchmarks `mch_bench_r19b08_icon1` and `mch_bench_r19b07_icon2`, we observe a speedup (on Balfrin) of about 1% compared to the previous test runs from 31.05.2023. When additionally including cuda-graphs, the code instead runs just as fast as the previous version tested on 31.05.2023.

For the benchmark `mch_bench_r19b08_kenda1`, we consistently ran into timeouts on Balfrin. This was caused by trying to fuse many nearby `PARALLEL` regions into one large `PARALLEL` region in `mo_latent_heat_nudging`. The reason for the slowdown was the line
```
DO jk = MINVAL(kqrs(i_startidx:i_endidx)), nlev
```
which led to the execution of a large `MINVAL` loop for each parallel thread. Combined with `GANG(STATIC: 1)`, this lead to large slowdowns. So we now added a local scalar `kqrs_min` which is simply set to `1` if OpenACC is defined, and to the `MINVAL` if OpenACC is not defined. The benchmark then runs about 5% faster than the previously tested version from 31.05.2023. And when additionally including cuda-graphs, the code instead runs just as fast as the previous version.

New merge request for icon/icon-nwp!869, since there were too many comments.

Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>