Fix validation of sstice_inst experiment
See merge request icon/icon-nwp!1031

The experiment `mch_opr_r04b07_sstice_inst` did not validate with probtest, which is why the tolerance test is currently deactivated on Balfrin (and on daint_gpu_nvidia_mixed).

We shorten the experiment from a simulation time of 2h to a simulation time of 60s, like most of the MeteoSwiss experiments.

This fixes the validation issues, so we reactivate the experiment for all builders on Balfrin and Daint.

Approved-by: Mikael Stellio <mikael.stellio@meteoswiss.ch>
Merged-by: Daniel Rieger <daniel.rieger@dwd.de>