[mpim] GPU port for derived output variables
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/209

This MR adds Harald Braun's work to enable a performant handling of derived output variables on GPUs.
    
The original work was tested by Marco Giorgetta and Claudia Frauen on the Levante-GPU partition based on the experiment `dy37_02_p1d_r2b7`, which produces time mean output fields. The same experiment has been used earlier for benchmarking on Daint, Juwels-Booster and Levante-CPU for the "ICON on GPU" article (Giorgetta et al., GMD, 2022).

The original work was carried out on the branch `icon-dkrz:hotfix-update-derived-quantities-on-gpu`.

Co-authored-by: Harald Braun <harald.braun@atos.net>
Merged-by: Monika Esch <monika.esch@mpimet.mpg.de>